requirements.txt
click==8.0.3
Flask==2.0.2
flask-cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
psycopg2-binary==2.9.3
Werkzeug==2.0.2
